refactor: winningNumbers class 파일로 분리
